Restructuring neuroml module to use submodule: standardmodels
