+++ pyNN/neuroml/standardmodels/synapses.py
@@ -1,0 +1,147 @@
+"""
+Standard synapses for the NeuroML module.
+
+:copyright: Copyright 2006-2017 by the PyNN team, see AUTHORS.
+:license: CeCILL, see LICENSE for details.
+"""
+
+from pyNN.standardmodels import synapses, build_translations
+from .simulator import state
+import logging
+
+import neuroml
+
+logger = logging.getLogger("PyNN_NeuroML")
+
+    
+
+class StaticSynapse(synapses.StaticSynapse):
+    __doc__ = synapses.StaticSynapse.__doc__
+    translations = build_translations(
+        ('weight', 'WEIGHT'),
+        ('delay', 'DELAY'),
+    )
+
+    def _get_minimum_delay(self):
+        d = state.min_delay
+        if d == 'auto':
+            d = state.dt
+        return d
+    
+    def add_to_nml_doc(self, nml_doc, population):
+        raise NotImplementedError()
+
+
+class TsodyksMarkramSynapse(synapses.TsodyksMarkramSynapse):
+    __doc__ = synapses.TsodyksMarkramSynapse.__doc__
+
+    translations = build_translations(
+        ('weight', 'WEIGHT'),
+        ('delay', 'DELAY'),
+        ('U', 'UU'),
+        ('tau_rec', 'TAU_REC'),
+        ('tau_facil', 'TAU_FACIL'),
+        ('u0', 'U0'),
+        ('x0', 'X' ),
+        ('y0', 'Y')
+    )
+    
+    def _get_minimum_delay(self):
+        d = state.min_delay
+        if d == 'auto':
+            d = state.dt
+        return d
+    
+    def add_to_nml_doc(self, nml_doc, population):
+        raise NotImplementedError()
+
+
+class STDPMechanism(synapses.STDPMechanism):
+    __doc__ = synapses.STDPMechanism.__doc__
+
+    base_translations = build_translations(
+        ('weight', 'WEIGHT'),
+        ('delay', 'DELAY')
+    )
+
+    def _get_minimum_delay(self):
+        d = state.min_delay
+        if d == 'auto':
+            d = state.dt
+        return d
+    
+    def add_to_nml_doc(self, nml_doc, population):
+        raise NotImplementedError()
+    
+
+class AdditiveWeightDependence(synapses.AdditiveWeightDependence):
+    __doc__ = synapses.AdditiveWeightDependence.__doc__
+
+    translations = build_translations(
+        ('w_max',     'wmax'),
+        ('w_min',     'wmin'),
+        ('A_plus',    'aLTP'),
+        ('A_minus',   'aLTD'),
+    )
+    
+    def add_to_nml_doc(self, nml_doc, population):
+        raise NotImplementedError()
+
+
+class MultiplicativeWeightDependence(synapses.MultiplicativeWeightDependence):
+    __doc__ = synapses.MultiplicativeWeightDependence.__doc__
+
+    translations = build_translations(
+        ('w_max',     'wmax'),
+        ('w_min',     'wmin'),
+        ('A_plus',    'aLTP'),
+        ('A_minus',   'aLTD'),
+    )
+    
+    def add_to_nml_doc(self, nml_doc, population):
+        raise NotImplementedError()
+
+
+class AdditivePotentiationMultiplicativeDepression(synapses.AdditivePotentiationMultiplicativeDepression):
+    __doc__ = synapses.AdditivePotentiationMultiplicativeDepression.__doc__
+
+    translations = build_translations(
+        ('w_max',     'wmax'),
+        ('w_min',     'wmin'),
+        ('A_plus',    'aLTP'),
+        ('A_minus',   'aLTD'),
+    )
+    
+    def add_to_nml_doc(self, nml_doc, population):
+        raise NotImplementedError()
+
+
+class GutigWeightDependence(synapses.GutigWeightDependence):
+    __doc__ = synapses.GutigWeightDependence.__doc__
+
+    translations = build_translations(
+        ('w_max',     'wmax'),
+        ('w_min',     'wmin'),
+        ('A_plus',    'aLTP'),
+        ('A_minus',   'aLTD'),
+        ('mu_plus',   'muLTP'),
+        ('mu_minus',  'muLTD'),
+    )
+    
+    def add_to_nml_doc(self, nml_doc, population):
+        raise NotImplementedError()
+
+
+class SpikePairRule(synapses.SpikePairRule):
+    __doc__ = synapses.SpikePairRule.__doc__
+
+    translations = build_translations(
+        ('tau_plus',  'tauLTP'),
+        ('tau_minus', 'tauLTD'),
+        ('A_plus',    'aLTP'),
+        ('A_minus',   'aLTD'),
+    )
+    
+    def add_to_nml_doc(self, nml_doc, population):
+        raise NotImplementedError()
+
